signup with useAuth hook. remove redux
