+++ src/containers/CreateBatch/Receipt.js
@@ -3,7 +3,6 @@
 /* eslint-disable no-unused-vars */
 
 import React, { useEffect, useState, useContext } from "react";
-import { connect } from "react-redux";
 import PropTypes from "prop-types";
 import { useHistory } from "react-router-dom";
 import { FormattedMessage } from "react-intl";
