--- src/containers/Login/actions.js
@@ -1,8 +1,0 @@
-import { SAVE_USER_DATA } from "./types";
-
-const saveUserAction = (userData) => ({
-  type: SAVE_USER_DATA,
-  payload: userData,
-});
-
-export default saveUserAction;
