+++ src/containers/Login/index.js
@@ -2,20 +2,17 @@
 import React, { useState, useEffect } from "react";
 import { useHistory, Link } from "react-router-dom";
 import { Button } from "antd";
-import PropTypes from "prop-types";
-import { connect } from "react-redux";
 
 import logo from "../../assets/bloinxLogo.png";
 import { validateEmail, validatePassword } from "./vlidators";
 import styles from "./index.module.scss";
-import saveUserAction from "./actions";
 import { useAuth } from "../../hooks/useAuth";
 
 const errors = {
   "auth/user-not-found": "El usuario no existe.",
 };
 
-function Login({ saveUser }) {
+function Login() {
   const history = useHistory();
   const [loading, setLoading] = useState(false);
   const [email, setEmail] = useState("");
@@ -40,7 +37,7 @@ function Login({ saveUser }) {
       userLogin: email,
       password,
       onSuccess: (data) => {
-        saveUser(data);
+        // saveUser(data);
         setLoading(false);
         history.push("/dashboard");
       },
@@ -141,18 +138,4 @@ function Login({ saveUser }) {
   );
 }
 
-Login.defaultProps = {
-  saveUser: () => {},
-};
-
-Login.propTypes = {
-  saveUser: PropTypes.func,
-};
-
-const mapStateToProps = (state) => state;
-
-const mapDispatchToProps = (dispatch) => ({
-  saveUser: (user) => dispatch(saveUserAction(user)),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(Login);
+export default Login;
