--- src/containers/Login/types.js
@@ -1,3 +1,0 @@
-export const SAVE_USER_DATA = "SAVE_USER_DATA";
-
-export default null;
