+++ src/index.js
@@ -2,11 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { IntlProvider } from "react-intl";
 import { BrowserRouter } from "react-router-dom";
-import { Provider } from "react-redux";
 
 import reportWebVitals from "./reportWebVitals";
 import flattenMessages from "./utils/locales";
-import store from "./redux/store";
 import es from "./locales/es.json";
 import en from "./locales/en.json";
 import App from "./App";
@@ -31,9 +29,7 @@ ReactDOM.render(
     >
       <BrowserRouter>
         <MainProvider>
-          <Provider store={store}>
-            <App />
-          </Provider>
+          <App />
         </MainProvider>
       </BrowserRouter>
     </IntlProvider>
