--- src/redux/reducers/rootReducer.js
@@ -1,10 +1,0 @@
-import { combineReducers } from "redux";
-import main from "./main";
-import userReducer from "../../containers/Login/reducer";
-
-const rootReducer = combineReducers({
-  main,
-  user: userReducer,
-});
-
-export default rootReducer;
