--- src/redux/types.js
@@ -1,5 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-
-export const INITIAL_CONTRACT_INSTANCE = "INITIAL_CONTRACT_INSTANCE";
-export const CURRENT_ADDRESS_WALLET = "CURRENT_ADDRESS_WALLET";
-export const CURRENT_PROVIDER = "CURRENT_PROVIDER";
